old-tests/submission/W3C/domhtml/HTMLOptionsCollection07.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - OPTIONSCOLLECTION</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/domcore/#dom-htmlcollection-item" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<form action="./files/getData.pl" id="form1" method="post" name="form1">
<p>
<select dir="ltr" id="selectId" multiple="multiple" name="select1" size="1" tabindex="7">
<option selected="selected" value="10001">EMP10001</option>
<option label="l1">EMP10002</option>
<option>EMP10003</option>
<option>EMP10004</option>
<option>EMP10005</option>
</select>
</p>
</form>
<p>
<select disabled="disabled" name="select2">
<option>EMP20001</option>
<option>EMP20002</option>
<option>EMP20003</option>
<option>EMP20004</option>
<option disabled="disabled">EMP20005</option>
</select>
</p>









<script>
var domtest = async_test("An HTMLOptionsCollection is a list of nodes representing HTML option     element.     An individual node may be accessed by either ordinal index, the node's     name or id attributes.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var optionsNode;
      var optionsValueNode;
      var optionsList;
      var vvalue;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("select");
      assert_equals(nodeList.length,2);
testNode = nodeList.item(0);
      optionsList = testNode.options;

      optionsNode = optionsList.item(0);
      optionsValueNode = optionsNode.firstChild;

      vvalue = optionsValueNode.nodeValue;

      assert_equals(vvalue,"EMP10001");
       
    domtest.done();
}));

</script>
<p>An HTMLOptionsCollection is a list of nodes representing HTML option
    element.
    An individual node may be accessed by either ordinal index, the node's
    name or id attributes.  (Test ordinal index=0).
    The item() method retrieves a node specified by ordinal index. Nodes
    are numbered in tree order.  The index origin is 0.

    Retrieve the first SELECT element.  Create a HTMLOptionsCollection.
    Retrieve the first item in the list and examine its firstChild's
    nodeValue.</p>
<div id="log"></div>
</body></html>
